+++ app/src/main/java/com/jorgedguezm/elections/data/room/Database.kt
@@ -2,9 +2,6 @@ package com.jorgedguezm.elections.data.room
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.jorgedguezm.elections.data.room.models.ElectionRaw
-import com.jorgedguezm.elections.data.room.models.PartyRaw
-import com.jorgedguezm.elections.data.room.models.ResultRaw
 
 @Database(entities = [ElectionRaw::class, PartyRaw::class, ResultRaw::class], version = 1,
     exportSchema = false)
